+++ test/unit/convertV2.test.js
@@ -110,7 +110,8 @@ const expect = require('chai').expect,
   multiExampleResponseCodeMatching =
     path.join(__dirname, VALID_OPENAPI_PATH, '/multiExampleResponseCodeMatching.json'),
   duplicateCollectionVars =
-    path.join(__dirname, VALID_OPENAPI_PATH, '/duplicateCollectionVars.json');
+    path.join(__dirname, VALID_OPENAPI_PATH, '/duplicateCollectionVars.json'),
+  issue795 = path.join(__dirname, VALID_OPENAPI_PATH, '/form-binary-file.json');
 
 
 describe('The convert v2 Function', function() {
@@ -2818,5 +2819,26 @@ describe('The convert v2 Function', function() {
         expect(conversionResult.output[0].data.variable[1]).to.have.property('key', 'MyParam');
         done();
       });
+  });
+
+  it('[Github #795] Should properly convert format binary to form data', function (done) {
+    var openapi = fs.readFileSync(issue795, 'utf8'),
+      reqBody, formData;
+    Converter.convertV2({ type: 'string', data: openapi }, {
+      requestNameSource: 'Fallback',
+      indentCharacter: 'Space',
+      collapseFolders: true,
+      optimizeConversion: true,
+      parametersResolution: 'schema'
+    }, (err, conversionResult) => {
+
+      reqBody = conversionResult.output[0].data.item[0].item[0].request.body;
+      formData = reqBody.formdata[0];
+
+      expect(err).to.be.null;
+      expect(conversionResult.result).to.equal(true);
+      expect(formData.type).to.equal('file');
+      done();
+    });
   });
 });